Add fail logic to the role
If no Transparent HugePage setting is provided the role will fail. This
is done because if no settings are provided an invalid unit file would
be generated. A molecule scenario is created to test this and both the
GitHub Actions and pre-commit configurations are updated to support
this new scenario.

Co-authored-by: Shane Frasier <jeremy.frasier@gwe.cisa.dhs.gov>