Add comments explaining why we are not using alphabetical order
When listing out valid values from settings they are not always listed
in alphabetical order. This is intentionally done so that it aligns
with what one sees when checking the current configuration of a setting
on a live system. Since we are deviating from our normal practices it
is important that we document the decision and include the rationale.

Co-authored-by: dav3r <david.redmin@gwe.cisa.dhs.gov>
Co-authored-by: Shane Frasier <jeremy.frasier@gwe.cisa.dhs.gov>